+++ src/components/App.tsx
@@ -8,13 +8,12 @@ import Portfolio from './portfolio/Portfolio';
 import Header from './header/Header';
 import Footer from './footer/Footer';
 import Profile from './shell/profile/Profile';
-// TODO: use @path variables in tsconfig
 import Imprint from '../pages/Imprint';
 import AboutUs from '../pages/AboutUs';
 import Privacy from '../pages/Privacy';
 import Page404 from '../pages/Page404';
 import Confirm from '../pages/Confirm';
-import PasswordReset from '../pages/PasswordReset';
+import Reset from '../pages/Reset';
 import { ContextProvider } from '../state/context';
 
 declare module '@material-ui/core/styles/createPalette' {
@@ -102,7 +101,7 @@ const App: React.FC = () => (
             <Privacy path="/privacy" />
             <Profile path="/profile" />
             <Confirm path="/confirm/:id/:token" />
-            <PasswordReset path="/passwordreset/:id/:token" />
+            <Reset path="/reset/:id/:token" />
             <Page404 default />
           </Router>
         </div>
