+++ src/components/shell/reset/ResetSuccessWindow.tsx
